Example with memfault added to subcomponent
Typically Memfault is added to the `main` top-level component, which
also has all the ESP-IDF common components included (some of which
Memfault has an implicit dependency on, which could be improved in the
Memfault SDK).

Alternate ESP-IDF application configurations include adding Memfault to
a project subcomponent.

This is an example of that setup, showcasing how Memfault is attached to
the subcomponent instead of the top-level application component.